Move DomainPicker List and Button to separate directories
