--- client/landing/gutenboarding/components/domain-picker/index.ts
@@ -1,1 +1,0 @@
-export { default as DomainPickerButton } from './button';
