+++ client/landing/gutenboarding/components/domain-picker/style.scss
@@ -1,15 +1,5 @@
 @import 'assets/stylesheets/gutenberg-base-styles';
 
-.domain-picker__button {
-	.dashicon {
-		margin-left: 0.5em;
-		transition: transform 100ms ease-in-out;
-	}
-	&.is-open .dashicon {
-		transform: rotate( 180deg );
-	}
-}
-
 .domain-picker__choose-domain-header {
 	font-weight: bold;
 	text-align: center;
